refactor(model): use iced registers and objectfile addresses
Python keys MIR, LIR and Addr on iced_x86 register numbers; the drafts'
PhysicalRegister numbering differed. Addr, Space and Group now live in
objectfile::module, ported from objectfile/module.py.

+++ src/model/memory.rs
@@ -8,7 +8,7 @@ use std::collections::BTreeSet;
 use std::fmt;
 
 use crate::model::mir::{Symbol, Value};
-use crate::old::object::omf::module::Space;
+use crate::objectfile::module::Space;
 
 /// Python `qbopt.model.memory:Kind`.
 #[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
@@ -127,7 +127,7 @@ pub enum ObjectIdentity {
     },
     TaggedIndex {
         tag: ObjectTag,
-        index: u32,
+        index: i64,
     },
     CallAllocation {
         callee: String,
